# golang_updates/drt-repos/drt-go-chain-proxy-1.2.0/cmd/proxy/config/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dharitri Gateway API",
    "description": "![Dharitri](/mvx-icon.png)\n\nThe documentation describes the endpoints that are available on the drt-go-chain-proxy project. \n\nGithub URL: `https://github.com/TerraDharitri/drt-go-chain-proxy`\n\nThis is used on `gateway.dharitri.org` while the public api (`api.dharitri.org`) is a wrapper over it.\n\nThis API is organized around REST principles, so if you've interacted with RESTful APIs before, many of the concepts will look familiar.\n\n## Request / Response Format\n\nJSON will be returned for all responses, including errors. Empty or blank fields are omitted. Requests with a message body use JSON as well. Successful requests will return a `2xx` HTTP status.\n\nThe general structure of a response is:\n``` JSON\n{\n  data  anything\n  error string\n  code  string\n}\n```\nWhere:\n- data: the expected result if the request was successful. `null` otherwise\n- error: if the request failed, the reason will be returned in this field. empty otherwise\n- code: the internal code for the request. can be `successful`, `internal_issue` or `bad_request`\n",
    "version": "1.1.0"
  },
  "paths": {
...
    "/node/old-storage-token/{token}/nonce/{nonce}": {
      "get": {
        "tags": [
          "node"
        ],
        "summary": "will return a boolean value indicating if the token's metadata is still stored in accounts' trie, as opposed to the system account",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "the token to look after",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nonce",
            "in": "path",
            "required": true,
            "description": "the nonce to look after",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/node/waiting-epochs-left/{key}": {
      "get": {
        "tags": [
          "node"
        ],
        "summary": "will return the number of epochs left for the public key until it becomes eligible",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "the public key to look after",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
...
    "/internal/{shard}/raw/block/by-nonce/{nonce}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the block at the given nonce following the protocol structure",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "nonce",
            "in": "path",
            "description": "the nonce to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/{shard}/raw/block/by-hash/{hash}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the block with the given hash following the protocol structure",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "description": "the hash to look after",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/{shard}/json/block/by-nonce/{nonce}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the block at the given nonce following the protocol structure, JSON encoded",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "nonce",
            "in": "path",
            "description": "the nonce to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/{shard}/json/block/by-hash/{hash}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the block with the given hash following the protocol structure, JSON encoded",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "description": "the hash to look after",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/{shard}/raw/miniblock/by-hash/{hash}/epoch/{epoch}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the miniblock with the given hash following the protocol structure",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "description": "the hash to look after",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "epoch",
            "in": "path",
            "description": "the epoch to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/{shard}/json/miniblock/by-hash/{hash}/epoch/{epoch}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the miniblock with the given hash following the protocol structure, JSON encoded",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "description": "the hash to look after",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "epoch",
            "in": "path",
            "description": "the epoch to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/raw/startofepoch/metablock/by-epoch/{epoch}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the epoch start block for a given epoch, following the protocol structure",
        "parameters": [
          {
            "name": "epoch",
            "in": "path",
            "description": "the epoch to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/json/startofepoch/metablock/by-epoch/{epoch}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the epoch start block for a given epoch, following the protocol structure, JSON encoded",
        "parameters": [
          {
            "name": "epoch",
            "in": "path",
            "description": "the epoch to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
...
  },
...
  "servers": [],
  "components": {
...
    "schemas": {
      "GenericResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
    }
  }
}